Added recursive code to generate deeper network
Created function in league.php to recursively generate as deep of a
network as the user wants, and generate a table so that I can see how
the variables change. getSummonerNetwork now sends back a json object
with two children: "graph_data" and "table_data". Graph_data consists of
the links and nodes for the FD graph, table_data is the HTML for the
variable table